requirements.txt
openpyxl
pandas
numpy
Pillow
piexif
